refactor(probeservices): merge orchestra.go and client.go
Part of https://github.com/ooni/probe-engine/issues/651
