--- probeservices/orchestra.go
@@ -1,24 +1,0 @@
-package probeservices
-
-import (
-	"errors"
-)
-
-var (
-	errInvalidMetadata = errors.New("invalid metadata")
-	errNotLoggedIn     = errors.New("not logged in")
-	errNotRegistered   = errors.New("not registered")
-)
-
-func (c Client) getCredsAndAuth() (*LoginCredentials, *LoginAuth, error) {
-	state := c.StateFile.Get()
-	creds := state.Credentials()
-	if creds == nil {
-		return nil, nil, errNotRegistered
-	}
-	auth := state.Auth()
-	if auth == nil {
-		return nil, nil, errNotLoggedIn
-	}
-	return creds, auth, nil
-}
